cloud_function_weather/requirements.txt
pandas
requests
python-dotenv
google-cloud-bigquery

